+++ confluence-client/src/main/kotlin/com/github/zeldigas/confclient/ConfluenceClientImpl.kt
@@ -18,12 +18,16 @@ import io.ktor.client.request.forms.*
 import io.ktor.client.statement.*
 import io.ktor.http.*
 import io.ktor.serialization.jackson.*
+import io.ktor.util.cio.*
+import io.ktor.utils.io.*
 import io.ktor.utils.io.streams.*
+import java.nio.file.Path
 import java.security.cert.X509Certificate
 import javax.net.ssl.X509TrustManager
 import kotlin.io.path.fileSize
 
 class ConfluenceClientImpl(
+    override val confluenceBaseUrl: Url,
     private val apiBase: String,
     private val httpClient: HttpClient
 ) : ConfluenceClient {
@@ -31,6 +35,9 @@ class ConfluenceClientImpl(
     companion object {
         private const val PAGE_SIZE = 100
     }
+
+    override val confluenceApiBaseUrl: Url
+        get() = Url(apiBase)
 
     override suspend fun describeSpace(key: String, expansions: List<String>): Space {
         return httpClient.get("$apiBase/space/$key") {
@@ -47,6 +54,15 @@ class ConfluenceClientImpl(
         val results = findPages(space, title, status, expansions)
 
         return extractSinglePage(results)
+    }
+
+    override suspend fun getPageById(
+        id: String,
+        expansions: Set<String>
+    ): ConfluencePage {
+        return httpClient.get("$apiBase/content/$id") {
+            addExpansions(expansions)
+        }.body()
     }
 
     override suspend fun getPageOrNull(
@@ -252,6 +268,14 @@ class ConfluenceClientImpl(
         httpClient.delete("$apiBase/content/$attachmentId").body<String>()
     }
 
+    override suspend fun downloadAttachment(attachment: Attachment, destination: Path) {
+        val downloadLink = attachment.links["download"] ?: throw IllegalArgumentException("No download link found: ${attachment.links}")
+
+        val response = httpClient.get(makeLink(confluenceBaseUrl, downloadLink))
+
+        response.bodyAsChannel().copyAndClose(destination.toFile().writeChannel())
+    }
+
     private fun FormBuilder.addAttachmentToForm(attachment: PageAttachmentInput) {
         append("comment", attachment.comment ?: "")
         append(
@@ -303,5 +327,5 @@ fun confluenceClient(
         }
     }
 
-    return ConfluenceClientImpl("${config.server}/rest/api", client)
+    return ConfluenceClientImpl(config.server, "${config.server}/rest/api", client)
 }