+++ convert/src/main/kotlin/com/github/zeldigas/text2confl/convert/markdown/export/ConfluenceLinksResolver.kt
@@ -1,0 +1,13 @@
+package com.github.zeldigas.text2confl.convert.markdown.export
+
+interface ConfluenceLinksResolver {
+
+    companion object {
+        val NOP = object : ConfluenceLinksResolver {
+            override fun resolve(space: String?, title: String): String = "wiki://$space/$title"
+        }
+    }
+
+    fun resolve(space: String?, title: String): String
+
+}